Added type cast to limit user controllable variable

+++ epubreader.php
@@ -15,7 +15,7 @@ require_once 'base.php';
 
 initURLParam();
 
-$idData = getURLParam('data', NULL);
+$idData = (int)getURLParam('data', NULL);
 $add = 'data=' . $idData . '&';
 if (!is_null (GetUrlParam (DB))) {
     $add .= DB . '=' . GetUrlParam (DB) . '&';
